+++ HISTORY.md
@@ -4,6 +4,8 @@ https://github.com/josdejong/mathjs
 
 ## not yet released, version 0.23.0
 
+- Renamed functions `edivide`, `emultiply`, and `epow` to `dotdivide`, 
+  `dotmultiply`, and `dotpow` respectively. 
 - Renamed function `unary` to `unaryminus` and added support for strings.
 - Implemented function `unaryplus` and unary plus operator.
 - Added constant `phi`, the golden ratio (`phi = 1.618...`).